+++ outputs.tf
@@ -55,21 +55,11 @@ output "worker_security_group_id" {
 }
 
 output "worker_iam_role_name" {
-  description = "IAM role name attached to the first EKS worker group"
-  value       = "${element(aws_iam_role.workers.*.name, 0)}"
+  description = "default IAM role name for EKS worker groups"
+  value       = "${element(aws_iam_role.workers.name, 0)}"
 }
 
 output "worker_iam_role_arn" {
-  description = "IAM role ID attached to the first EKS worker group"
-  value       = "${element(aws_iam_role.workers.*.arn, 0)}"
-}
-
-output "worker_iam_role_names" {
-  description = "IAM role names attached to EKS workers"
-  value       = "${aws_iam_role.workers.*.name}"
-}
-
-output "worker_iam_role_arns" {
-  description = "IAM role IDs attached to EKS workers"
-  value       = "${aws_iam_role.workers.*.arn}"
+  description = "default IAM role name for EKS worker groups"
+  value       = "${element(aws_iam_role.workers.arn, 0)}"
 }